[modify] add io and visualization in utils
